+++ gulpfile.js
@@ -8,6 +8,8 @@ var webpack = require('webpack');
 var fs = require('fs');
 var clear = require('clear');
 var plumber = require('gulp-plumber');
+var BundleAnalyzerPlugin = require('webpack-bundle-analyzer')
+  .BundleAnalyzerPlugin;
 
 gulp.task('lint:src', function lintSrc() {
   return gulp
@@ -48,49 +50,68 @@ gulp.task(
 gulp.task(
   'build:browser',
   gulp.series('lint:src', function buildBrowser() {
-    return (
-      gulp
-        .src('src/browser.js')
-        .pipe(plumber())
-        .pipe(
-          plugins.webpack({
-            output: { library: 'StellarSdk' },
-            module: {
-              loaders: [
-                {
-                  test: /\.js$/,
-                  exclude: /node_modules/,
-                  loader: 'babel-loader'
-                },
-                { test: /\.json$/, loader: 'json-loader' }
-              ]
-            },
-            plugins: [
-              // Ignore native modules (ed25519)
-              new webpack.IgnorePlugin(/ed25519/)
+    return gulp
+      .src('src/browser.js')
+      .pipe(plumber())
+      .pipe(
+        plugins.webpack({
+          output: { library: 'StellarSdk' },
+          module: {
+            loaders: [
+              {
+                test: /\.js$/,
+                exclude: /node_modules/,
+                loader: 'babel-loader'
+              },
+              { test: /\.json$/, loader: 'json-loader' }
             ]
-          })
-        )
-        // Add EventSource polyfill for IE11
-        .pipe(
-          plugins.insert.prepend(
-            fs.readFileSync(
-              './node_modules/event-source-polyfill/src/eventsource.min.js'
-            )
-          )
-        )
-        .pipe(plugins.rename('stellar-sdk.js'))
-        .pipe(gulp.dest('dist'))
-        .pipe(
-          plugins.uglify({
-            output: {
-              ascii_only: true
-            }
-          })
-        )
-        .pipe(plugins.rename('stellar-sdk.min.js'))
-        .pipe(gulp.dest('dist'))
-    );
+          },
+          plugins: [
+            // Ignore native modules (ed25519)
+            new webpack.IgnorePlugin(/ed25519/)
+          ]
+        })
+      )
+      .pipe(plugins.rename('stellar-sdk.js'))
+      .pipe(gulp.dest('dist'))
+      .pipe(
+        plugins.uglify({
+          output: {
+            ascii_only: true
+          }
+        })
+      )
+      .pipe(plugins.rename('stellar-sdk.min.js'))
+      .pipe(gulp.dest('dist'));
+  })
+);
+
+gulp.task(
+  'analyze:browser',
+  gulp.series('lint:src', function analyzeBrowser() {
+    return gulp
+      .src('src/browser.js')
+      .pipe(plumber())
+      .pipe(
+        plugins.webpack({
+          output: { library: 'StellarSdk' },
+          module: {
+            loaders: [
+              {
+                test: /\.js$/,
+                exclude: /node_modules/,
+                loader: 'babel-loader'
+              },
+              { test: /\.json$/, loader: 'json-loader' }
+            ]
+          },
+          plugins: [
+            // Ignore native modules (ed25519)
+            new webpack.IgnorePlugin(/ed25519/),
+            new BundleAnalyzerPlugin()
+          ]
+        })
+      );
   })
 );