+++ test/integration/server_test.js
@@ -128,6 +128,9 @@ describe('integration tests', function() {
             expect(operation.account).to.equal(randomAccount.publicKey());
             eventStreamClose();
             done();
+          },
+          onerror: (err) => {
+            done(err);
           }
         });