+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 As this project is pre 1.0, breaking changes may happen for minor version bumps.
 A breaking change will get clearly marked in this log.
 
+## In master
+
+- Replace the eventsource polyfill with one that doesn't use http/https, which
+  fixes a bug with Angular 6
+
 ## [v0.14.0](https://github.com/stellar/js-stellar-sdk/compare/v0.13.0...v0.14.0)
 
 - Updated some out-of-date dependencies