+++ CHANGELOG.md
@@ -9,9 +9,8 @@ A breaking change will get clearly marked in this log.
   polyfill. If you plan to support IE11 / Edge, please use
   [`event-source-polyfill`](https://www.npmjs.com/package/event-source-polyfill)
   to set `window.EventSource`.
-- Fix a bug with Angular 6
-  - Upgrade `stellar-base` to a version that doesn't use the `crypto` library
-  - Replace the eventsource polyfill with one that doesn't use http/https
+- Upgrade `stellar-base` to a version that doesn't use the `crypto` library,
+  fixing a bug with Angular 6
 
 ## [v0.14.0](https://github.com/stellar/js-stellar-sdk/compare/v0.13.0...v0.14.0)