+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 As this project is pre 1.0, breaking changes may happen for minor version bumps.
 A breaking change will get clearly marked in this log.
 
+## In master
+
+- **Breaking change**: `stellar-sdk` no longer ships with an `EventSource`
+  polyfill. If you plan to support IE11 / Edge, please use
+  [`event-source-polyfill`](https://www.npmjs.com/package/event-source-polyfill)
+  to set `window.EventSource`.
+- Upgrade `stellar-base` to a version that doesn't use the `crypto` library,
+  fixing a bug with Angular 6
+
 ## [v0.14.0](https://github.com/stellar/js-stellar-sdk/compare/v0.13.0...v0.14.0)
 
 - Updated some out-of-date dependencies